active_recon/requirements.txt
arpreq==0.3.3
ipaddress==1.0.23
netifaces==0.10.9
python-nmap==0.6.1
pyroute2==0.5.7

